Add support for netlist "statics".
a.k.a. Verilog parameters.
a.k.a. VHDL generics.

Add support in GenVerilog.

TODO add support for GenVHDL.  GenVHDL works right now, it just ignores the
statics.

TODO support in Netlist for assigning statics when instantiating a module

TODO bump the .cabal versions up when this work is finished.