verilog: Tweak Expression.hs
Rewrite Expression to Expression' and parametrize it over the type for literals.
Define type Expression = Expression' Number.  This allows the user to
instantiate Expression' with a more structured type for literals, so we can
define a simulator or compiler more easily.

No changes necessary outside of this module, since the new definition of
Expression and ConstExpr are identical to before.